% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCSList.R
\name{BuildCSList}
\alias{BuildCSList}
\title{Build a comma separated List}
\usage{
BuildCSList(data, pri, sec)
}
\arguments{
\item{data}{data frame containing primary and secondary data columns}

\item{pri}{Primary field name (repeating values)}

\item{sec}{Secondary field (values would be added to same record,
comma separated)}
}
\value{
a data frame with two fields Primary and secondary (comma
separated list)
}
\description{
Builds a comma separated list given a data frame with two fiends, primary
repeating values and secondary values to be summarized with comma separated
in the same field
}
\examples{
scnames <- c("Abrothrix longipilis", "Abrothrix jelskii")
SynList <- GetITISSyn(scnames)
BuildCSList(SynList,"Name","Syn")

}
\concept{List functions}
