% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_scientific.R
\name{check_scientific}
\alias{check_scientific}
\title{Parse and resolve a scientific name string}
\usage{
check_scientific(name)
}
\arguments{
\item{name}{scientific name string to be checked}
}
\value{
Resolved canonical name (NULL if not matched)
}
\description{
Parse the name using GNR and GBIF parse API to make sure the name
is scientific name
}
\examples{
check_scientific("Akodon longipilis (Waterhouse, 1837)")
check_scientific("Mus longipilis Waterhouse, 1837")
check_scientific("Akodon hershkovitzi Patterson, Gallardo, and Freas, 1984")

}
\seealso{
Other Name functions: 
\code{\link{cast_canonical}()},
\code{\link{expand_name}()},
\code{\link{get_accepted_names}()},
\code{\link{guess_taxo_level}()},
\code{\link{list_higher_taxo}()},
\code{\link{melt_canonical}()}
}
\concept{Name functions}
