% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckScientificName.R
\name{CheckScientificName}
\alias{CheckScientificName}
\title{Parse and resolve a scientific name string}
\usage{
CheckScientificName(name)
}
\arguments{
\item{name}{scientific name string to be checked}
}
\value{
Resolved Canonical name and NULL is not matched
}
\description{
Parse the names using GBIF parse API to make sure the name passed
is an scientific name
}
\examples{
CheckScientificName("Akodon longipilis (Waterhouse, 1837)")
CheckScientificName("Mus longipilis Waterhouse, 1837")
CheckScientificName("Akodon hershkovitzi Patterson, Gallardo, and Freas, 1984")

}
\seealso{
Other Name functions: \code{\link{ExpandSyn}},
  \code{\link{GuessTaxoLevel}}, \code{\link{expand_name}}
}
\concept{Name functions}
