% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwC2taxo.R
\name{DwC2taxo}
\alias{DwC2taxo}
\title{Darwin Core to Taxolist format}
\usage{
DwC2taxo(namelist)
}
\arguments{
\item{namelist}{names list in Darwin Core format}
}
\value{
names list is taxolist format
}
\description{
Converts a Darwin Core name list to taxolist format
}
\details{
The name lists downloaded for ITIS website in Darwin Core format has
all the required fields. Just needs to be converted and quality checked in terms
 of missing linkages
}
\examples{
\dontrun{
if(interactive()){
 taxolist <- DwC2taxo(namelist)
 }
}
}
