% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpandSyn.R
\name{ExpandSyn}
\alias{ExpandSyn}
\title{Expands Scientific name is the genus is used with '.'}
\usage{
ExpandSyn(name, syn)
}
\arguments{
\item{name}{scientific name}

\item{syn}{synonym with short form genus name to expand the Genus}
}
\value{
Synonym with Genus expanded using either name or previous names in
the syn list
}
\description{
This function is depreciated. Please use \link{expand_name}
}
\examples{
\dontrun{
ExpandSyn("Addax gibbosa", "A. mytilopes")
ExpandSyn("Oryx addax", "O. nasomaculatus")
}
}
\seealso{
Other Discontinued functions: \code{\link{BuildCSList}},
  \code{\link{CheckScientificName}},
  \code{\link{GetWikiSyn}}, \code{\link{GuessTaxoLevel}},
  \code{\link{MakeCanonical}}, \code{\link{MatchLists}},
  \code{\link{make_canonical}}
}
\concept{Discontinued functions}
