% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_itis_syn.R
\name{get_itis_syn}
\alias{get_itis_syn}
\title{Get Integrated Taxonomic Information System 'ITIS' Synonyms for a Scientific
 Name}
\usage{
get_itis_syn(scname)
}
\arguments{
\item{scname}{Scientific Name}
}
\value{
a list containing synonyms
}
\description{
Fetch Synonyms using Integrated Taxonomic Information System 'ITIS' web
 service
}
\examples{
\donttest{
get_itis_syn("Abrothrix longipilis")
get_itis_syn("Abditomys latidens")
}
}
\seealso{
Other ITIS functions: 
\code{\link{list_itis_syn}()}
}
\concept{ITIS functions}
