% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetITISSyn.R
\name{GetITISSyn}
\alias{GetITISSyn}
\title{Get ITIS Synonyms for list of names}
\usage{
GetITISSyn(namelist)
}
\arguments{
\item{namelist}{list of scientific names}
}
\value{
a data frame containing names (passed) and synonyms
}
\description{
This function is depreciated. Please use \link{list_itis_syn}
}
\details{
Fetch Synonyms from ITIS
}
\examples{
#GetITISSyn("Abrothrix longipilis")
#GetITISSyn(c("Abditomys latidens", "Abeomelomys sevia", "Abrothrix jelskii" ))

}
\seealso{
Other ITIS functions: \code{\link{ListITISSyn}},
  \code{\link{get_itis_syn}}, \code{\link{list_itis_syn}}
}
\concept{ITIS functions}
