% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessTaxoLevel.R
\name{GuessTaxoLevel}
\alias{GuessTaxoLevel}
\title{Guess the level of Scientific Name}
\usage{
GuessTaxoLevel(name)
}
\arguments{
\item{name}{scientific name string to be checked}
}
\value{
Guess on level of Taxon name and NULL if not sure
}
\description{
#' This function is depreciated. Please use \link{match_lists}
}
\examples{
\dontrun{
GuessTaxoLevel("Akodon longipilis")
GuessTaxoLevel("Akodon")
GuessTaxoLevel("Abrocoma cinerea shistacea")
}
}
\seealso{
Other Discontinued functions: \code{\link{BuildCSList}},
  \code{\link{CheckScientificName}},
  \code{\link{ExpandSyn}}, \code{\link{GetWikiSyn}},
  \code{\link{MakeCanonical}}, \code{\link{MatchLists}}
}
\concept{Discontinued functions}
