% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_itis_syn.R
\name{list_itis_syn}
\alias{list_itis_syn}
\title{Get ITIS Synonyms for list of names}
\usage{
list_itis_syn(namelist)
}
\arguments{
\item{namelist}{list of scientific names}
}
\value{
a data frame containing canonical names (passed) and synonyms
}
\description{
Fetch Synonyms from Integrated Taxonomic Information System 'ITIS'
}
\examples{
\dontrun{
list_itis_syn("Abrothrix longipilis")
list_itis_syn(c("Abditomys latidens", "Abeomelomys sevia", "Abrothrix jelskii" ))
}
}
\seealso{
Other ITIS functions: 
\code{\link{get_itis_syn}()}
}
\concept{ITIS functions}
