% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_wiki_syn.R
\name{list_wiki_syn}
\alias{list_wiki_syn}
\title{Get Wikipedia Synonyms for list of names}
\usage{
list_wiki_syn(namelist, verbose = TRUE)
}
\arguments{
\item{namelist}{list of scientific names}

\item{verbose}{status output. Default TRUE}
}
\value{
a data frame containing names, synonyms and Canonical synonyms matched
   with is scientific name backbone taxonomy  
   \describe{ \item{Name}{: Scientific name}
   \item{WikiName}{: Wikipedia page name}
   \item{OrigSyn}{: Original synonym returned by Wikipedia}
 \item{Syn}{: Synonym in canonical form, matched with GBIF}}
}
\description{
Fetch Synonyms from Wikipedia and clean them for use
}
\examples{
\donttest{
 list_wiki_syn("Abrothrix illutea")
 list_wiki_syn(c("Abditomys latidens", "Abeomelomys sevia",
                 "Abrocoma schistacea"))
}

}
\concept{Wiki functions}
