% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_canonical.R
\name{make_canonical}
\alias{make_canonical}
\title{Construct canonical names}
\usage{
make_canonical(dat, genus = "", species = "", subspecies = "")
}
\arguments{
\item{dat}{data frame containing taxonomic list}

\item{genus}{field name for Genus field}

\item{species}{field name for Species field}

\item{subspecies}{field name for Subspecies field}
}
\value{
a data frame containing Canonical names field added or repopulated using
    filed names for Genus, Species and Subspecies specified in parameters
}
\description{
Construct canonical names using Genus, Species and Subspecies fields. At times
due to spaces or NAs in the data fields, it makes it tricky to generate
canonical names.
}
\examples{
\dontrun{
make_canonical(mylist,"genus","species","subspecies")
}
}
\concept{name functions}
