% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCanonical.R
\name{MakeCanonical}
\alias{MakeCanonical}
\title{Construct cononical names}
\usage{
MakeCanonical(dat, genus = "", species = "", subspecies = "")
}
\arguments{
\item{dat}{data frame containing taxonomic list}

\item{genus}{field name for Genus field}

\item{species}{field name for Species field}

\item{subspecies}{field name for Subspecies field}
}
\value{
a data frame containing Canonical names field added or repopulated using
    filed names for Genus, Species and Subspecies specified in paramaters
}
\description{
Cononical names using Genus, Species and Subspecies friends
}
\examples{
\dontrun{
MakeCanonical(mylist,"genus","species","subspecies")
}
}
\concept{name functions}
