% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_lists.R
\name{match_lists}
\alias{match_lists}
\title{match two taxonomic lists}
\usage{
match_lists(master, checklist, masterfld, checklistfld)
}
\arguments{
\item{master}{master taxonomic list}

\item{checklist}{match  taxonomic list}

\item{masterfld}{field name for canonical name in master list}

\item{checklistfld}{field name for canonical name in match list}
}
\value{
a list with two data frames containing matched and non-matched
names from the master list
}
\description{
match two taxonomic lists using canonical names
}
\examples{
\dontrun{
match_lists(master,checklist,"canonical","canonical")
}
}
\concept{list functions}
