% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_cs_field.R
\name{melt_cs_field}
\alias{melt_cs_field}
\title{Build a list seperating character (comma) separated field into seperate records}
\usage{
melt_cs_field(data, pri, sepchar = ",")
}
\arguments{
\item{data}{data frame containing a data columns with character(comma) seperated values}

\item{pri}{Primary field name with character(comma) seperated values}

\item{sepchar}{Character separator between the data items. Default is comma}
}
\value{
a data frame with seperate records for each value in primary field specified
}
\description{
Builds a list seperating character (comma) separated field given a data frame
with a field with repeating values
}
\examples{
\dontrun{
scnames <- c("Abrothrix longipilis", "Abrothrix jelskii")
syn_list <- list_itis_syn(scnames)
cs_syn_list <- cast_cs_field(syn_list ,"Name","Syn")
syn_list_new <- melt_cs_field(cs_syn_list,"Syn")
}

}
\seealso{
Other List functions: \code{\link{cast_cs_field}}
}
\concept{List functions}
