% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymize_subspecies.R
\name{synonymize_subspecies}
\alias{synonymize_subspecies}
\title{Convert all subspecies into synonyms of the species}
\usage{
synonymize_subspecies(master)
}
\arguments{
\item{master}{List of names with a field named canonical}
}
\value{
Same list of names with id and accid fields added (or data updated
the fields exists) with all subspecies linked to the species names as
synonyms
}
\description{
used in generating master lists
}
\details{
used in generating master lists
}
\examples{
\dontrun{
newmaster <- synosynonymize_subspecies(master)
}
}
