% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxotools-package.R
\docType{package}
\name{taxotools-package}
\alias{taxotools}
\alias{taxotools-package}
\title{taxotools: Taxonomic List Processing}
\description{
Taxonomic lists matching and merging, casting and melting scientific names, managing taxonomic lists from Global Biodiversity Information Facility 'GBIF' or Integrated Taxonomic Information System 'ITIS', harvesting names from Wikipedia and fuzzy matching.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/vijaybarve/taxotools/issues}
}

}
\author{
\strong{Maintainer}: Vijay Barve \email{vijay.barve@gmail.com} (\href{https://orcid.org/0000-0002-4852-2567}{ORCID})

}
\keyword{internal}
