% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxotools.R
\docType{package}
\name{taxotools}
\alias{taxotools}
\alias{taxotools-package}
\title{taxotools: Tools to handle taxonomic data}
\description{
Some tools to work with lists taxonomic names.
}
\section{List functions}{

\itemize{
\item \link{cast_cs_field}
\item \link{match_lists}
}
}

\section{Name functions}{

\itemize{
\item \link{expand_name}
\item \link{CheckScientificName}
\item \link{GuessTaxoLevel}
\item \link{MakeCanonical}
}
}

\section{ITIS functions}{

\itemize{
\item \link{GetITISSyn}
\item \link{ListITISSyn}
}
}

\section{Wiki functions}{

\itemize{
\item \link{GetWikiSyn}
}

#'@section Discontinued functions:
\itemize{
\item \link{MatchLists}
\item \link{BuildCSList}
\item \link{ExpandSyn}
}
}

\section{Citation}{

\itemize{
\item Barve, V., (2018). taxotools: Tools to handle
taxonomic data (R package V 0.0.1). Retrieved from
https://cran.r-project.org/web/packages/taxotools/index.html
}
}

