% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki2taxo.R
\name{wiki2taxo}
\alias{wiki2taxo}
\title{Wikipedia list to taxo}
\usage{
wiki2taxo(wikisyn)
}
\arguments{
\item{wikisyn}{Wikipedia synonyms list}
}
\value{
taxolist
}
\description{
Converts the output of \link{list_wiki_syn} function to
taxolist format of \link{taxotools} package
}
\details{
Output of \link{list_wiki_syn} function has different format
than taxolist. This function converts it making sure to add additional
fields and maitian the synonym linckages.
}
\examples{
\dontrun{
if(interactive()){
 wikilist <- list_wiki_syn("Abrothrix illutea")
 wiki2taxo(wikilist)
 }
}
}
\seealso{
Other List functions: 
\code{\link{cast_cs_field}()},
\code{\link{melt_cs_field}()},
\code{\link{merge_lists}()},
\code{\link{syn2taxo}()}
}
\concept{List functions}
