% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCSList.R
\name{BuildCSList}
\alias{BuildCSList}
\title{Build a comma separated List}
\usage{
BuildCSList(data, pri, sec, duplicate = FALSE, sepchar = ",")
}
\arguments{
\item{data}{data frame containing primary and secondary data columns}

\item{pri}{Primary field name (repeating values)}

\item{sec}{Secondary field (values would be added to same record,
comma separated)}

\item{duplicate}{If true, duplicate entries are allowed in secondary field}

\item{sepchar}{Character separator between the data items. Default is comma}
}
\value{
a data frame with two fields Primary and secondary (comma
separated list)
}
\description{
This function is depreciated. Please use \link{cast_cs_field}
}
\details{
Builds a comma separated list given a data frame with two fiends, primary
repeating values and secondary values to be summarized with comma separated
in the same field
}
\examples{
\dontrun{
scnames <- c("Abrothrix longipilis", "Abrothrix jelskii")
SynList <- GetITISSyn(scnames)
BuildCSList(SynList,"Name","Syn")
}

}
\seealso{
Other Discontinued functions: \code{\link{CheckScientificName}},
  \code{\link{ExpandSyn}}, \code{\link{GetWikiSyn}},
  \code{\link{GuessTaxoLevel}},
  \code{\link{MakeCanonical}}, \code{\link{MatchLists}}
}
\concept{Discontinued functions}
