% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckScientificName.R
\name{CheckScientificName}
\alias{CheckScientificName}
\title{Parse and resolve a scientific name string}
\usage{
CheckScientificName(name)
}
\arguments{
\item{name}{scientific name string to be checked}
}
\value{
Resolved Canonical name and NULL is not matched
}
\description{
This function is depreciated. Please use \link{check_scientific}
}
\details{
Parse the names using GBIF parse API to make sure the name passed
is an scientific name
}
\examples{
\dontrun{
CheckScientificName("Akodon longipilis (Waterhouse, 1837)")
CheckScientificName("Mus longipilis Waterhouse, 1837")
CheckScientificName("Akodon hershkovitzi Patterson, Gallardo, and Freas, 1984")
}
}
\seealso{
Other Discontinued functions: \code{\link{BuildCSList}},
  \code{\link{ExpandSyn}}, \code{\link{GuessTaxoLevel}},
  \code{\link{MatchLists}}
}
\concept{Discontinued functions}
