% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_name.R
\name{expand_name}
\alias{expand_name}
\title{Expands Scientific name}
\usage{
expand_name(fullname, shortname)
}
\arguments{
\item{fullname}{full scientific name}

\item{shortname}{scientific name with short form genus name to expand the Genus}
}
\value{
scientific name with Genus expanded using reference name provided as parameter
}
\description{
At times the genus is specified with first character and '.' rather then
repeating genus names every time. These are either synonyms or species of the
same genus listed one bellow another. To convert these names to canonical
names, we need to expand the genus name (typically) using previous entry in
the list.
}
\examples{
expand_name("Addax gibbosa", "A. mytilopes")
expand_name("Oryx addax", "O. nasomaculatus")

}
\seealso{
Other Name functions: 
\code{\link{build_gen_syn}()},
\code{\link{cast_canonical}()},
\code{\link{cast_scientificname}()},
\code{\link{check_scientific}()},
\code{\link{get_accepted_names}()},
\code{\link{guess_taxo_rank}()},
\code{\link{list_higher_taxo}()},
\code{\link{melt_canonical}()},
\code{\link{melt_scientificname}()},
\code{\link{resolve_names}()},
\code{\link{taxo_fuzzy_match}()}
}
\concept{Name functions}
