% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetITISSyn.R
\name{GetITISSyn}
\alias{GetITISSyn}
\title{Get ITIS Synonyms for list of names}
\usage{
GetITISSyn(namelist)
}
\arguments{
\item{namelist}{list of scientific names}
}
\value{
a data frame containing names (passed) and synonyms
}
\description{
Fetch Synonyms from ITIS
}
\seealso{
Other ITIS functions: \code{\link{ListITISSyn}}
}
\concept{ITIS functions}
