% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWikiSyn.R
\name{GetWikiSyn}
\alias{GetWikiSyn}
\title{Get Wikipedia Synonyms for list of names}
\usage{
GetWikiSyn(namelist)
}
\arguments{
\item{namelist}{list of scientific names}
}
\value{
a data frame containing names, synonyms and Canonical synonyms matched
   with GBIF backbone taxonomy  \itemize{ \item{Name}  {: Scientific name}
   \item{OrigSyn}  {: Original synonym returned by Wikipedia}
 \item{Syn}  {: Synonym in canonical form, matched with GBIF}}
}
\description{
This function is depreciated. Please use \link{list_wiki_syn}
}
\details{
Fetch Synonyms from Wikipedia and clean them for use
}
\examples{
#GetWikiSyn("Abrothrix illutea")
#GetWikiSyn(c("Abditomys latidens", "Abeomelomys sevia", "Abrocoma schistacea"))

}
\seealso{
Other Discontinued functions: \code{\link{BuildCSList}},
  \code{\link{CheckScientificName}},
  \code{\link{ExpandSyn}}, \code{\link{GuessTaxoLevel}},
  \code{\link{MakeCanonical}}, \code{\link{MatchLists}},
  \code{\link{make_canonical}}
}
\concept{Discontinued functions}
