% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessTaxoLevel.R
\name{GuessTaxoLevel}
\alias{GuessTaxoLevel}
\title{Guess the level of Scientific Name}
\usage{
GuessTaxoLevel(name)
}
\arguments{
\item{name}{scientific name string to be checked}
}
\value{
Guess on level of Taxon name and NULL if not sure
}
\description{
Guess the level of Scientific Name
}
\seealso{
Other Name functions: \code{\link{CheckScientificName}},
  \code{\link{ExpandSyn}}
}
\concept{Name functions}
