% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListITISSyn.R
\name{ListITISSyn}
\alias{ListITISSyn}
\title{Get list ITIS Synonyms for a Scientific Name}
\usage{
ListITISSyn(scname)
}
\arguments{
\item{scname}{Scientific Name}
}
\value{
a list containing synonyms
}
\description{
Fetch Synonyms using ITIS web service
}
\examples{
ListITISSyn("Abrothrix longipilis")
ListITISSyn("Abditomys latidens")

}
\seealso{
Other ITIS functions: \code{\link{GetITISSyn}}
}
\concept{ITIS functions}
