% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchLists.R
\name{MatchLists}
\alias{MatchLists}
\title{match two taxonomic lists}
\usage{
MatchLists(master, lst, masterfld, lstfld)
}
\arguments{
\item{master}{master taxonomic list}

\item{lst}{match  taxonomic list}

\item{masterfld}{field name for canonical name in master list}

\item{lstfld}{field name for canonical name in match list}
}
\value{
a list with two data frames containing matched and non-matched names from the master list
}
\description{
This function is depreciated. Please use \link{match_lists}
}
\details{
match two taxonomic lists using canonical names
}
\examples{
\dontrun{
MatchLists(master,lst,"canonical","canonical")
}
}
\seealso{
Other Discontinued functions: \code{\link{BuildCSList}},
  \code{\link{CheckScientificName}},
  \code{\link{ExpandSyn}}, \code{\link{GetWikiSyn}},
  \code{\link{GuessTaxoLevel}},
  \code{\link{MakeCanonical}}, \code{\link{make_canonical}}
}
\concept{Discontinued functions}
