% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_lists.R
\name{merge_lists}
\alias{merge_lists}
\title{merge two lists of names}
\usage{
merge_lists(master = NULL, checklist = NULL, verbose = TRUE)
}
\arguments{
\item{master}{master list of names}

\item{checklist}{list to be merged}

\item{verbose}{verbose output on the console}
}
\value{
returns three components. First the names to be added, second
the names that could not be matched and third the names that matched
multiple names in master
}
\description{
Useful in generating a master list of names form multiple
sources
}
\details{
Matches names is checklist with names on master
}
\examples{
\dontrun{
merge_lists(master = NA,
            checklist = NA,
            verbose = TRUE)
}
}
\seealso{
Other List functions: 
\code{\link{DwC2taxo}()},
\code{\link{cast_cs_field}()},
\code{\link{match_lists}()},
\code{\link{melt_cs_field}()},
\code{\link{syn2taxo}()},
\code{\link{synonymize_subspecies}()},
\code{\link{taxo2DwC}()},
\code{\link{taxo2syn}()},
\code{\link{wiki2taxo}()}
}
\concept{List functions}
