% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn2taxo.R
\name{syn2taxo}
\alias{syn2taxo}
\title{Synonym list to taxolist}
\usage{
syn2taxo(synlist, canonical = "canonical", synonym = "synonym")
}
\arguments{
\item{synlist}{Synonym list with Accepted name (canonical) and Synonym columns}

\item{canonical}{Accepted names column name, Default: 'canonical'}

\item{synonym}{Synonym column name , Default: 'synonym'}
}
\value{
returns a data frame in taxolist format  with all the names in 
same columns and accepted names linked to synonyms using id and accid fields
}
\description{
Converts a Synonym list with Accepted Names and Synonym columns
to taxolist format
}
\details{
Converts a synonyms list to taxolist format
}
\examples{
\dontrun{
 synlist <- data.frame("id" = c(1,2,3),
                       "canonical" = c("Hypochlorosis ancharia",
                                       "Hypochlorosis ancharia",
                                       "Hypochlorosis ancharia"),
                       "synonym" = c( "Hypochlorosis tenebrosa",
                                      "Pseudonotis humboldti",
                                      "Myrina ancharia"),
                       "family" = c("Lycaenidae", "Lycaenidae", 
                                    "Lycaenidae"),
                       "source" = c("itis","wiki","wiki"),
                       stringsAsFactors = F)
                       
 mytaxo <- syn2taxo(synlist)
}
}
\seealso{
Other List functions: 
\code{\link{DwC2taxo}()},
\code{\link{cast_cs_field}()},
\code{\link{match_lists}()},
\code{\link{melt_cs_field}()},
\code{\link{merge_lists}()},
\code{\link{synonymize_subspecies}()},
\code{\link{taxo2DwC}()},
\code{\link{taxo2doc}()},
\code{\link{taxo2syn}()},
\code{\link{wiki2taxo}()}
}
\concept{List functions}
