% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxotools.R
\docType{package}
\name{taxotools}
\alias{taxotools}
\alias{taxotools-package}
\title{taxotools: Tools to handle taxonomic data}
\description{
Some tools to work with lists taxonomic names.
}
\section{List functions}{

\itemize{
\item \link{cast_cs_field}
\item \link{melt_cs_field}
\item \link{match_lists}
}
}

\section{Name functions}{

\itemize{
\item \link{expand_name}
\item \link{check_scientific}
\item \link{guess_taxo_level}
\item \link{make_canonical}
\item \link{list_higher_taxo}
}
}

\section{ITIS functions}{

\itemize{
\item \link{get_itis_syn}
#'\item \link{list_itis_syn}
}
}

\section{Wiki functions}{

\itemize{
\item \link{list_wiki_syn}
}
}

\section{Discontinued functions}{

\itemize{
\item \link{MatchLists}
\item \link{BuildCSList}
\item \link{ExpandSyn}
\item \link{CheckScientificName}
\item \link{GuessTaxoLevel}
\item \link{MakeCanonical}
\item \link{GetWikiSyn}
\item \link{GetITISSyn}
\item \link{ListITISSyn}
}
}

\section{Citation}{

\itemize{
\item Barve, V., (2019). taxotools: Tools to handle
taxonomic data (R package V 0.0.20). Retrieved from
https://cran.r-project.org/web/packages/taxotools/index.html
}
}

